Revert "[Test-Data] Move data to sub-folder"
This reverts commit 5622336ec323aabfdef4628e22ad33b87edbd156.
