--- admin/src/test-data/data/spaces.ts
@@ -1,55 +1,0 @@
-import { Space } from '@phits-tech/common/dist/dao-firestore'
-
-export const spaces: Space[] = [
-  {
-    slug: 'maker-club-nu',
-    name: 'Maker Club, NU',
-    description: 'Aliquid quod tenetur neque labore sequi non facere. Consequatur voluptatibus quae quisquam explicabo odio, hic natus dolorem, alias error, corporis totam eveniet.',
-    banner: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Fmaker-club-nu-banner.jpg?alt=media&token=9ccefce1-b0d2-4808-a4a3-109ed37a9194',
-    logo: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Fmaker-club-nu-logo.jpg?alt=media&token=da3b874d-ba62-4bd9-b236-43cab188a51c',
-    category: 'community',
-    locationText: '4th floor SC2, Naresuan University',
-    locationLatitude: 16.742_235_085_922_57,
-    locationLongitude: 100.193_572_537_988_8
-  },
-  {
-    slug: 'setic-nu',
-    name: 'SET Investment Center, NU',
-    description: 'Nostrum laborum nemo odio obcaecati omnis amet at pariatur soluta, accusamus incidunt aliquid nesciunt cupiditate quasi cumque beatae nihil perspiciatis ad exercitationem!',
-    banner: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Fsetic-nu-banner.jpg?alt=media&token=a5ce286e-37d1-4cbe-80ed-e14db07c4438',
-    logo: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Fsetic-nu-logo.jpg?alt=media&token=66a5c132-087f-4ac4-a34d-5a40e68962d9',
-    category: 'coworking',
-    locationText: 'Near BEC, Naresuan University',
-    locationLatitude: 16.748_156_316_592_908,
-    locationLongitude: 100.196_996_297_037_84
-  },
-  {
-    slug: 'usit-nu',
-    name: 'USIT, NU',
-    description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolorem commodi dolor corrupti et a vitae aliquam, minima ipsa at quasi. Delectus et ad quis eos repellendus illo, ducimus labore enim!',
-    banner: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Fusit-nu-banner.jpg?alt=media&token=4a5b2039-a513-48c3-9d26-719ddc8d97c2',
-    logo: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Fusit-nu-logo.jpg?alt=media&token=c82966c5-28a3-4dd5-bdfe-5719257a7ddd',
-    category: 'cafe',
-    locationText: 'Near Engineering, Naresuan University',
-    locationLatitude: 16.744_032_818_990_863,
-    locationLongitude: 100.195_160_148_854_64
-  },
-  {
-    slug: 'pt-discord',
-    name: 'PT Discord',
-    description: 'Optio esse fugit voluptatum! Nesciunt, similique quaerat distinctio voluptatem voluptates facilis totam ex repudiandae. Incidunt quaerat eligendi consequatur non labore facere nulla.',
-    banner: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Fmodern-technology.jpeg?alt=media&token=e440f586-b143-4ce0-9039-a1e03921e441',
-    logo: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Fphits-tech-logo.png?alt=media&token=a3e436e1-f303-4083-b3d6-e324f4073e3c',
-    category: 'online',
-    locationText: 'https://discord.gg/4CHb53F'
-  },
-  {
-    slug: 'pt-youtube',
-    name: 'PT YouTube',
-    description: 'Facilis rem quas perspiciatis quis quae numquam blanditiis ea sed recusandae aspernatur inventore reprehenderit ut, ullam, asperiores itaque officia perferendis officiis incidunt!',
-    banner: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Ftarget-marketing.jpeg?alt=media&token=aec127f6-e685-40de-a09c-416450b0dcf3',
-    logo: 'https://firebasestorage.googleapis.com/v0/b/phits-tech-emu.appspot.com/o/spaces%2Fphits-tech-logo.png?alt=media&token=a3e436e1-f303-4083-b3d6-e324f4073e3c',
-    category: 'online',
-    locationText: 'https://www.youtube.com/channel/UC83wlCFD2-FOYxSUzVG-oUg'
-  }
-]
